Refactor plot_task into separate plot.py module
Moved plot_task function from utils.py to a dedicated plot.py file to
improve code organization and separate plotting functionality from
utility functions. Updated imports in main.py accordingly.
