+++ utils.py
@@ -1,6 +1,3 @@
-import matplotlib.pyplot as plt
-from matplotlib.colors import ListedColormap, Normalize
-
 from random import choice, randint, sample, shuffle, uniform
 
 from dsl import *
@@ -91,34 +88,6 @@ def format_task(
         'train': [format_example(example) for example in task['train']],
         'test': [format_example(example) for example in task['test']]
     }
-
-
-def plot_task(
-    task: List[dict],
-    title: str = None
-) -> None:
-    """
-    displays a task
-    """
-    cmap = ListedColormap([
-        '#000', '#0074D9', '#FF4136', '#2ECC40', '#FFDC00',
-        '#AAAAAA', '#F012BE', '#FF851B', '#7FDBFF', '#870C25'
-    ])
-    norm = Normalize(vmin=0, vmax=9)
-    args = {'cmap': cmap, 'norm': norm}
-    height = 2
-    width = len(task)
-    figure_size = (width * 3, height * 3)
-    figure, axes = plt.subplots(height, width, figsize=figure_size)
-    for column, example in enumerate(task):
-        axes[0, column].imshow(example['input'], **args)
-        axes[1, column].imshow(example['output'], **args)
-        axes[0, column].axis('off')
-        axes[1, column].axis('off')
-    if title is not None:
-        figure.suptitle(title, fontsize=20)
-    plt.subplots_adjust(wspace=0.1, hspace=0.1)
-    plt.show()
 
 
 def fix_bugs(
